requirements.txt
Flask==2.0.1
requests==2.26.0
websocket-client==1.2.1
